Custom domains for AFQTS
Create custom domains for AFQTS for each environment. Use the domains/environment
terraform module and add a redirect ruleset.

This is based on the front-door "Standard" offering which is not fully
ready in terraform. See progress in: https://github.com/hashicorp/terraform-provider-azurerm/pull/16671

The rest of the set-up is made manually for now and will be added here
once it's ready in terraform.