Add logstash pipeline for TRA
Logit has a pipeline that processes incoming log messages and extracts
data into separate, queryable fields (accessed via Settings > Logstash
Pipelines).

This is changed via a free-form editor in the Logit UI. A version
history is kept by Logit but prior versions must be manually requested
from their customer services team as and when required.

To encourage some level of review around changes to the pipeline, we're
going to add it to this repo.

For now, simply add the pipeline implementation at time of writing,
making no changes.