Add Azure-specific section to Logstash pipeline
The existing pipeline file is written with GOVUK PaaS apps in mind,
and references various fields specific to a PaaS deployment. The recent
crop of services deployed to Azure (Refer misconduct, Access teaching
qualifications, Check teacher record) send log messages in a different
format and with different fields. They share a common logging approach
(similar libraries and set up), in which the key difference from the
PaaS apps is that the message payload is in JSON.

This commit encloses all of the existing code in a conditional check,
specifically for the presence of "space_name" in the incoming log
message. If this field isn't present, we assume the log message is from
Azure and process it differently. Specifically, we parse the message
field of the incoming event as JSON, and write its data to a set of keys
prefixed with "azure_app".

This preserves all existing behaviour while giving us queryable fields
for the services sending their logs in JSON format.